+++ tests/example.py
@@ -20,8 +20,8 @@ class WorkerSignalQuit(Worker):
     """
     max_concurrent_tasks = 1
 
-    async def schedule_job(self, queue, data):
-        await super().schedule_job(queue, data)
+    async def schedule_job(self, *args):
+        await super().schedule_job(*args)
         if self.jobs_complete >= 2:
             self.handle_sig(2, None)
 
@@ -30,7 +30,7 @@ class WorkerSignalTwiceQuit(Worker):
     """
     worker which simulates receiving sigint twice after 2 jobs
     """
-    async def schedule_job(self, queue, data):
-        await super().schedule_job(queue, data)
+    async def schedule_job(self, *args):
+        await super().schedule_job(*args)
         if self.jobs_complete >= 2:
             self.handle_sig_force(2, None)