+++ tests/test_cli.py
@@ -45,15 +45,34 @@ def test_worker_exited_badly(tmpworkdir, monkeypatch):
 def test_main_process_sigint(tmpworkdir, monkeypatch, caplog):
     monkeypatch.setattr(arq.worker.Process, 'start', MagicMock())
     monkeypatch.setattr(arq.worker.Process, 'join', MagicMock())
-    is_alive = MagicMock(side_effect=[True, True, False, False])
-    monkeypatch.setattr(arq.worker.Process, 'is_alive', is_alive)
+    monkeypatch.setattr(arq.worker.Process, 'is_alive', MagicMock(return_value=True))
     monkeypatch.setattr(arq.worker.Process, 'exitcode', 0)
     monkeypatch.setattr(arq.worker.Process, 'pid', 123)
+
+    os_kill = MagicMock()
+    monkeypatch.setattr(arq.worker.os, 'kill', os_kill)
+
     tmpworkdir.join('test.py').write(EXAMPLE_FILE)
     work_runner = arq.worker.RunWorkerProcess('test.py', 'Worker')
     work_runner.handle_sig(signal.SIGINT, None)
-    assert is_alive.call_count == 3
     assert 'got signal: SIGINT, waiting for worker pid=123 to finish...' in caplog
+    os_kill.assert_called_once_with(123, arq.worker.SIG_SUPERVISOR)
+
+
+def test_main_process_sigint_worker_stopped(tmpworkdir, monkeypatch, caplog):
+    monkeypatch.setattr(arq.worker.Process, 'start', MagicMock())
+    monkeypatch.setattr(arq.worker.Process, 'join', MagicMock())
+    monkeypatch.setattr(arq.worker.Process, 'is_alive', MagicMock(return_value=False))
+    monkeypatch.setattr(arq.worker.Process, 'exitcode', 0)
+    monkeypatch.setattr(arq.worker.Process, 'pid', 123)
+
+    os_kill = MagicMock()
+    monkeypatch.setattr(arq.worker.os, 'kill', os_kill)
+
+    tmpworkdir.join('test.py').write(EXAMPLE_FILE)
+    work_runner = arq.worker.RunWorkerProcess('test.py', 'Worker')
+    work_runner.handle_sig(signal.SIGINT, None)
+    assert os_kill.called is False
 
 
 def test_main_process_sigint_twice(tmpworkdir, monkeypatch, caplog):
@@ -88,6 +107,5 @@ def test_main_process_sigint_twice_worker_running(tmpworkdir, monkeypatch, caplo
     with pytest.raises(arq.worker.ImmediateExit):
         work_runner.handle_sig_force(signal.SIGINT, None)
     assert is_alive.call_count == 1
-    assert os_kill.called
-    assert os_kill.call_args == ((123, signal.SIGTERM),)
+    os_kill.assert_called_once_with(123, signal.SIGTERM)
     assert 'got signal: SIGINT again, forcing exit' in caplog